% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{AmpFileRead}
\alias{AmpFileRead}
\title{Load amplifier amplitude file}
\usage{
AmpFileRead(FILENAME, ChannelNumber = 32L)
}
\arguments{
\item{FILENAME}{A string as path to file with single "/".}

\item{ChannelNumber}{An integer indicating number of channels in recording (default = 32).}
}
\value{
Returns an armadillo matrix (rows = ChannelNumber, columns = time).
}
\description{
This function extracts data points from a binary file (amplifier.dat) generated by the intan system.
}
