% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{arma_apply_median}
\alias{arma_apply_median}
\title{Median Spike}
\usage{
arma_apply_median(input_mat)
}
\arguments{
\item{input_mat}{An input matrix with amplitudes for single spikes.}
}
\value{
Returns a vector with length of concatenated samples.
}
\description{
This function uses an amplitude matrix from single spikes in channel.
}
