% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{arma_spike_extraction_loop}
\alias{arma_spike_extraction_loop}
\title{Spike extraction loop}
\usage{
arma_spike_extraction_loop(input_mat_raw, channel_nr)
}
\arguments{
\item{input_mat_raw}{A numeric vector from a matrix with amplitudes for single spikes.}

\item{channel_nr}{An int indicating the number of recorded channels.}
}
\value{
Returns a numeric spike vector with corrected baseline for single channels.
}
\description{
This function uses an amplitude matrix from single spikes in channel to output baseline corrected median spikes.
}
