% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{BaselineZScore}
\alias{BaselineZScore}
\title{Baseline Z-scoring

This function computes the phase lock using the Rho vector length 
(strength of the locking) and the circular mean phase from a list containing 
wavelet transforms of epochs. The input list has to be flattend (vectorised) 
first.}
\usage{
BaselineZScore(InputMat, from, to)
}
\arguments{
\item{InputMat}{A matrix with time on the x-axis.}

\item{from}{A starting point for baseline used for z-score.}

\item{to}{An end point for baseline used for z-score.}
}
\value{
Returns a matrix with z-scores based on baseline as reference.
}
\description{
Baseline Z-scoring

This function computes the phase lock using the Rho vector length 
(strength of the locking) and the circular mean phase from a list containing 
wavelet transforms of epochs. The input list has to be flattend (vectorised) 
first.
}
