% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{BinaryFileAccess}
\alias{BinaryFileAccess}
\title{Binary file access}
\usage{
BinaryFileAccess(
  FILENAME,
  spikePoints,
  WINDOW = 40L,
  CHANNELCOUNT = 32L,
  CACHESIZE = 51200L,
  BYTECODE = 2L
)
}
\arguments{
\item{FILENAME}{A string as path to file with single "/".}

\item{spikePoints}{An integer vector for time points which correspond to sampling points.}

\item{WINDOW}{An integer indicating points taken before and after time point (default = 40).}

\item{CHANNELCOUNT}{An integer indicating number of channels in recording (default = 32).}

\item{CACHESIZE}{An integer indicating the size of cache used for buffering file in bytes (default = 512000).}

\item{BYTECODE}{An integer indicating number of bytes coding for bit rate (8bit = 1, 16bit = 2 etc.) (default = 2).}
}
\value{
Returns an armadillo cube with extracted spikes.
}
\description{
This function extracts data points from a binary file based on time points and a window
which will cut out the data of and concatenates them based on the channels.
}
