% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{BWFiltCpp}
\alias{BWFiltCpp}
\title{Butterworth filter c++}
\usage{
BWFiltCpp(
  InputSignal,
  SamplingFrequency,
  ORDER = 2L,
  f0 = 1000,
  type = "low",
  BatchSize = 40000L,
  CORES = 1L
)
}
\arguments{
\item{InputSignal}{A complex matrix from FFTW.}

\item{SamplingFrequency}{A double indicating sampling frequency.}

\item{ORDER}{An int as filtering order (default = 2).}

\item{f0}{A double as cutoff frequency (default = 10).}

\item{type}{A string indicating the filter type ("low", "high"). The default is "low".}

\item{BatchSize}{An integer indicating the batch size for FFT (default = 4e4).}

\item{CORES}{An int indicating the number of threads used (default = 1).}
}
\value{
Filtered signal as numeric vector.
}
\description{
This function returns a filtered Signal.
}
