% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwFilter.R
\name{bwFilter}
\alias{bwFilter}
\title{Butterworth filter}
\usage{
bwFilter(
  Signal,
  order = 2,
  SamplingFrequency = 20000,
  f0,
  cores = 1,
  type = "zero",
  PaddingType
)
}
\arguments{
\item{Signal}{A Numeric Vector as input}

\item{order}{An Integer indicating the order of filtering (default = 2)}

\item{SamplingFrequency}{An Integer as sampling frequency of signal (default = 2e4)}

\item{f0}{The input for the frequency cutoff}

\item{cores}{Multicore usages can be activated by setting number of threads (default = 1).}

\item{type}{Type of filtering can be set as "low", "high" for low-pass and high-pass filtering.}

\item{PaddingType}{Padding type can be set by as "zero" (zero-padding), "edge" (values of edge are used for padding), "none" (no padding). The default value is set "zero"}
}
\value{

}
\description{
This function filters the data based on the butterworth filter. It uses an FFTW implementation to accelerate the filtering process.
}
