% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{BWFilterCpp}
\alias{BWFilterCpp}
\title{Butterworth filter}
\usage{
BWFilterCpp(
  InputFFT,
  SamplingFrequency,
  ORDER = 2L,
  f0 = 10,
  type = "low",
  CORES = 1L
)
}
\arguments{
\item{InputFFT}{A complex matrix from FFTW.}

\item{SamplingFrequency}{A double indicating sampling frequency.}

\item{ORDER}{An int as filtering order (default = 2).}

\item{f0}{A double as cutoff frequency.}

\item{type}{A string indicating the filter type ("low", "high"). The default is "low".}

\item{CORES}{An int indicating the number of threads used (default = 1).}
}
\value{
Complex armadillo column vector.
}
\description{
This function returns a modified FFT.
}
