% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{chan_out}
\alias{chan_out}
\title{Maximum amplitude channel}
\usage{
chan_out(spike_shape_list, channel_nr)
}
\arguments{
\item{spike_shape_list}{A numeric vector from a matrix with amplitudes for single spikes.}

\item{channel_nr}{An int indicating the number of recorded channels.}
}
\value{
Returns a numeric matrix inlcuding the channel number and amplitude.
}
\description{
This function takes a list of single events and computes the channel with max amplitude and amplitude for all units in list.
}
