% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ConfIntPoisson}
\alias{ConfIntPoisson}
\title{Confidence interval for poisson train}
\usage{
ConfIntPoisson(
  CountVector,
  CONFLEVEL = 0.95,
  SD = 0.6,
  CENTREMIN = 0.6,
  KERNELSIZE = 20L
)
}
\arguments{
\item{CountVector}{An integer vector of counts estimated by Spike CCF.}

\item{CONFLEVEL}{A double indicating the confidence-level (default = 0.95).}

\item{SD}{A double as standard deviation for a gaussian shape parameter (default = 0.6).}

\item{CENTREMIN}{A double as shape parameter determining the strength of centre exclusion.}

\item{KERNELSIZE}{A double as length parameter for gaussian kernel (2*KERNELSIZE+1).}
}
\value{
Returns a list containing counts per bin, axis, random bin count, confidence-intervals with counts per bin.
}
\description{
Computes the confidence interval of a poisson train using an inverted gaussian
distribution. Based on the weighted average with the given distribution lamda
can be estimated. The confidence interval is then estimated by estimation through
Chisq-distribution.
}
