% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConfIntPoisson.R
\name{ConfIntPoissonR}
\alias{ConfIntPoissonR}
\title{Confidence interval poisson}
\usage{
ConfIntPoissonR(
  CountVector,
  conf.level = 0.95,
  SD = 0.6,
  CentreMin = 0.6,
  KernelSize = 65
)
}
\arguments{
\item{CountVector}{An integer vector of counts}

\item{conf.level}{A double indication the confidence interval.}

\item{SD}{Parameter of inverted gaussian.}

\item{CentreMin}{Parameter of inverted gaussian.}

\item{KernelSize}{Parameter of inverted gaussian.}
}
\value{

}
\description{
This function computes a confidence interval for a series of counts.
The data is convolved with a inverted gaussian to calculate a weighted mean as lamda.
From the new lamda the confidence interval is computed using the chisq distribution as approximation.
}
