% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{convertToBinary}
\alias{convertToBinary}
\title{Convert digital inputs.}
\usage{
convertToBinary(StimTrace)
}
\arguments{
\item{StimTrace}{vector.}
}
\value{
Returns a list with the matrix, the channel ID and the label "Digital".
}
\description{
This function converts digital inputs of the intan system and separates the channels into a matrix.
}
