% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CxCubeCollapse}
\alias{CxCubeCollapse}
\title{Average complex matrix (from wavelet power cube)}
\usage{
CxCubeCollapse(x)
}
\arguments{
\item{x}{A cube with complex matrices each slice representing ERP.}
}
\value{
An average complex matrix.
}
\description{
This function computes the average complex matrix of a complex cube.
}
