% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{decimate}
\alias{decimate}
\title{Decimate}
\usage{
decimate(SIGNAL, FIR_FILTER, M)
}
\arguments{
\item{SIGNAL}{A numeric vector with length N.}

\item{FIR_FILTER}{A predefined FIR filter as numeric vector (low-pass filter for <= target frequency).}

\item{M}{An int representing downsampling factor.}
}
\value{
Returns a numeric vector with length N/M.
}
\description{
This function downsamples a signal using a preset FIR filter which should be filtering at half the target frequency.
}
