% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ERPMat}
\alias{ERPMat}
\title{ERP (event-related potential) extraction

This function returns a matrix with extracted traces for any given range. 
The range is has to be provided in form of onset indeces and end indeces. 
Additional information can be provided to adjust for sampling differences 
between range points and trace sampling frequency. If required the window 
is proportionally elongated to include Pre/Post times.}
\usage{
ERPMat(
  Trace,
  Onset,
  End,
  SamplingFreqStim = 0,
  SamplingFreqTrace = 0,
  PrePhase = FALSE,
  PostPhase = FALSE
)
}
\arguments{
\item{Trace}{A numeric vector which is used for extraction.}

\item{Onset}{An integer vector with onset sample indeces.}

\item{End}{An integer vector with end sample indeces.}

\item{SamplingFreqStim}{A double indicating the sampling frequency of onset/end indeces.}

\item{SamplingFreqTrace}{A double indicating the sampling frequency of the trace.}

\item{PrePhase}{A bool indicating if pre-onset timings should be included (default = false).}

\item{PostPhase}{A bool indicating if post-end timings should be included (default = false).}
}
\value{
Returns a matrix with extracted ERPs.
}
\description{
ERP (event-related potential) extraction

This function returns a matrix with extracted traces for any given range. 
The range is has to be provided in form of onset indeces and end indeces. 
Additional information can be provided to adjust for sampling differences 
between range points and trace sampling frequency. If required the window 
is proportionally elongated to include Pre/Post times.
}
