% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{FiringRate}
\alias{FiringRate}
\title{ERP (event-related potential) extraction}
\usage{
FiringRate(
  SpikeTimes,
  timeStart = NULL,
  timeEnd = NULL,
  sigma = NULL,
  alpha = NULL,
  useBAKS = FALSE,
  BAKSalpha = 4,
  BAKSbeta = 4,
  SamplingRate = 1000,
  CORES = 4L
)
}
\arguments{
\item{SpikeTimes}{A numeric vector with spike times in seconds.}

\item{timeStart}{A double for the start time for which the rate should be computed.}

\item{timeEnd}{A double for the end time for which the rate should be computed.}

\item{sigma}{A double indicating sigma of the gausian kernel in seconds.}

\item{alpha}{A double indicating alpha of the gamma kernel in seconds.}

\item{useBAKS}{A indicating whether Bayesian Adaptive Kernel Smoother should be used.}

\item{BAKSalpha}{A double indicating the prior for alpha in BAKS (default = 4).}

\item{BAKSbeta}{A double indicating the prior for beta in BAKS (default = 4).}

\item{SamplingRate}{A double indicating the sampling rate (default = 1000).}

\item{CORES}{An integer indicating how many core should be used (default = 4).}
}
\value{
Returns a vector with firing rate.
}
\description{
This function returns a matrix with extracted traces for any given range. 
The range is has to be provided in form of onset indeces and end indeces. 
Additional information can be provided to adjust for sampling differences 
between range points and trace sampling frequency. If required the window 
is proportionally elongated to include Pre/Post times.
}
