% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{FiringRateSparse}
\alias{FiringRateSparse}
\title{Firing Rate of Neuron (as sparse vector)

This function computes the firing rate based on convolution with gaussian kernel and the 
convolution with gamma distribution. The output is a sparse vector which returns the rate 
for the position of the SpikeTime.}
\usage{
FiringRateSparse(SpikeTimes, sigma = NULL, alpha = NULL, SamplingRate = 1000)
}
\arguments{
\item{SpikeTimes}{A numeric vector with spike times in seconds.}

\item{sigma}{A double indicating sigma of the gausian kernel in seconds.}

\item{alpha}{A double indicating alpha of the gamma kernel in seconds.}

\item{SamplingRate}{A double indicating the sampling rate (default = 1000).}
}
\value{
Returns a vector with firing rate.
}
\description{
Firing Rate of Neuron (as sparse vector)

This function computes the firing rate based on convolution with gaussian kernel and the 
convolution with gamma distribution. The output is a sparse vector which returns the rate 
for the position of the SpikeTime.
}
