% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetChanMap.R
\name{GetChanMap}
\alias{GetChanMap}
\title{Get channel map}
\usage{
GetChanMap(directory, FileName = "chanMap")
}
\arguments{
\item{directory}{String to input directory.}

\item{FileName}{Name of file without ".mat".}
}
\value{
Returns a data frame containing the coordinates of the channels.
}
\description{
Function wrapper to load channel map from a given directory.
}
