% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDirectories.R
\name{getDataDirectory}
\alias{getDataDirectory}
\title{Get Data Directory}
\usage{
getDataDirectory(
  mainDirectory = NULL,
  recLabels = NULL,
  stimLabels = NULL,
  stimRecIdentical = NULL
)
}
\arguments{
\item{mainDirectory}{A string locating the path.}

\item{recLabels}{A vector of strings including the recording label names.}

\item{stimLabels}{A vector of strings including the stimulation label names.}

\item{stimRecIdentical}{A bool indicating whether stimulation and recording directory are identical.}
}
\value{
Returns a data.frame which includes the labels and the directory to the folders.
}
\description{
This function will get the directory for any given label by opening a user interaction.
}
