% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDirectories.R
\name{getDirectory}
\alias{getDirectory}
\title{Get Directory}
\usage{
getDirectory(mainDirectory, Labels)
}
\arguments{
\item{mainDirectory}{A string locating the path.}

\item{Labels}{A vector of strings including the label names.}
}
\value{
Returns a data.frame which includes the labels and the directory to the folders.
}
\description{
This function will get the directory for any given label by opening a user interaction.
}
