\name{InVivoR-package}
\alias{InVivoR-package}
\alias{InVivoR}
\docType{package}
\title{\packageTitle{InVivoR}}
\description{\packageDescription{InVivoR}}
\section{Package Content}{\packageIndices{InVivoR}}
\author{\packageAuthor{InVivoR}}
\section{Maintainer}{\packageMaintainer{InVivoR}}
\keyword{package}
