% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetChanMap.R
\name{loadChanMap}
\alias{loadChanMap}
\title{Load channel map}
\usage{
loadChanMap(FileName)
}
\arguments{
\item{FileName}{Complete filename including path and file ending.}
}
\value{
Returns a data frame containing the coordinates of the channels.
}
\description{
Function to load channel.
}
