% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mat_baseline_zscore}
\alias{mat_baseline_zscore}
\title{Phase lock analysis}
\usage{
mat_baseline_zscore(input_mat, from_point, to_point)
}
\arguments{
\item{input_mat}{A matrix with time on the x-axis.}

\item{from_point}{A starting point for baseline used for z-score.}

\item{to_point}{An end point for baseline used for z-score.}
}
\value{
Returns a matrix with z-scores based on baseline as reference.
}
\description{
This function computes the phase lock using the Rho vector length 
(strength of the locking) and the circular mean phase from a list containing 
wavelet transforms of epochs. The input list has to be flattend (vectorised) 
first.
}
