% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{max_channel}
\alias{max_channel}
\title{Maximum Amplitude Channel}
\usage{
max_channel(median_input_mat)
}
\arguments{
\item{median_input_mat}{A median matrix from arma_apply_median.}
}
\value{
Returns a vector with position of maximum amplitude and amplitude itself.
}
\description{
This function uses the median spike shape to find channel with maximum amplitude.
As input the function requires the previously computed median spike matrix with dimensions of n.row = clusters and n.col = channel count
}
