% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{MaxChannel}
\alias{MaxChannel}
\title{Maximum amplitude channel

This function uses the median spike shape to find channel with maximum amplitude.
As input the function requires the previously computed median spike matrix with dimensions of n.row = clusters and n.col = channel count}
\usage{
MaxChannel(MedianSpikeMat)
}
\arguments{
\item{MedianSpikeMat}{A median matrix from arma_apply_median.}
}
\value{
Returns a vector with position of maximum amplitude and amplitude itself.
}
\description{
Maximum amplitude channel

This function uses the median spike shape to find channel with maximum amplitude.
As input the function requires the previously computed median spike matrix with dimensions of n.row = clusters and n.col = channel count
}
