% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{MI}
\alias{MI}
\title{Modulation index analysis

This function returns a list including a modulation index matrix, the phase frequency and the power frequency.
there are leading NA, marking the leadings NA as TRUE and
everything else as FALSE.}
\usage{
MI(
  PowerMatRaw,
  PhaseMatRaw,
  PowerPeriods,
  PhasePeriods,
  BIN_NUMBER,
  PHASE_FREQ_START,
  PHASE_FREQ_END,
  POWER_FREQ_START,
  POWER_FREQ_END,
  CORES
)
}
\arguments{
\item{PowerMatRaw}{A matrix with power.}

\item{PhaseMatRaw}{A matrix with phase.}

\item{PowerPeriods}{A double vector with unique periods (1/Frequency) for power matrix.}

\item{PhasePeriods}{A double vector with unique periods (1/Frequency) for phase matrix.}

\item{BIN_NUMBER}{An integer indicating in how many parts phase should be split.}

\item{PHASE_FREQ_START}{A double indicating starting frequency (1/period) for phase.}

\item{PHASE_FREQ_END}{A double indicating end frequency (1/period) for phase.}

\item{POWER_FREQ_START}{A double indicating starting frequency (1/period) for power.}

\item{POWER_FREQ_END}{A A double indicating end frequency (1/period) for power.}

\item{CORES}{An integer indicating number of threads used (multicore support).}
}
\value{
Returns a list containing the modulation index matrix and vectors containing its axis (Powerfrequency and Phasefrequency)
}
\description{
Modulation index analysis

This function returns a list including a modulation index matrix, the phase frequency and the power frequency.
there are leading NA, marking the leadings NA as TRUE and
everything else as FALSE.
}
