% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MITable.R
\name{MITable}
\alias{MITable}
\title{Table for modulation index}
\usage{
MITable(MIList, GapFill)
}
\arguments{
\item{MIList}{List generated by MI_analysis() which includes necessary modulation indeces.}

\item{GapFill}{double indicating the frequency gap in seconds (inverse of frequency).}
}
\value{

}
\description{
A function to compute a data.table of the modulation index.
}
