% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{morletWavlet}
\alias{morletWavlet}
\title{Morlet Wwavelet (time domain)}
\usage{
morletWavlet(t, sigma)
}
\arguments{
\item{t}{A numeric sequence of time (-t/2 to t/2) with steps of sampling frequency.}

\item{sigma}{A double indicating the shape parameter of the wavelet.}
}
\value{
Morlet wavelet as complex vector.
}
\description{
This function returns a complex morlet wavelet in the time domain. It can be 
used in convolution.
}
