% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{morletWT}
\alias{morletWT}
\title{Application of morlet wavelet (frequency domain)}
\arguments{
\item{SignalFFT}{A complex vector of the signal FFT.}

\item{scale}{A double indicating the scale parameter of daughter wavelet.}

\item{morletFFT}{A vector of wavelet in frequency domain.}

\item{LNorm}{A double indicating the L normalisation (power of 1/LNorm, default = 2).}
}
\value{
Morlet wavelet as numeric vector.
}
\description{
This function returns the convolution of a complex morlet daughter wavelet with the signal.
}
