% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{PhaseListAnalysis}
\alias{PhaseListAnalysis}
\title{Phase lock analysis}
\usage{
PhaseListAnalysis(x, DIM_X, DIM_Y, DIM_Z, CORES = 1L)
}
\arguments{
\item{x}{A numeric vector of phases (radians).}

\item{DIM_X}{An int for x-dimension of wavelet transform.}

\item{DIM_Y}{An int for y-dimension of wavelet transform.}

\item{DIM_Z}{An int for z-dimension of list (number of epochs).}

\item{CORES}{An int indicating the number of threads used (default = 1).}
}
\value{
Returns a list containing a matrix with Rho vector lengths and a matrix with the corresponding circular mean.
}
\description{
This function computes the phase lock using the Rho vector length 
(strength of the locking) and the circular mean phase from a list containing 
wavelet transforms of epochs. The input list has to be flattend (vectorised) 
first.
}
