% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{PhaseListAnalysisResample}
\alias{PhaseListAnalysisResample}
\title{Matrix resample}
\usage{
PhaseListAnalysisResample(x, SHUFFLES = 200L, CORES = 1L)
}
\arguments{
\item{x}{A matrix.}

\item{SHUFFLES}{An int indicating the number of shuffles.}

\item{CORES}{An int indicating the number of threads used (default = 1).}
}
\value{
Returns a matrix indicating the probability of value being larger than shuffled data.
}
\description{
This function shuffles the data independently of dimensions with 
resampling from a input matrix. It returns the probability for a any 
given value in a matrix to be larger than a random sample.
}
