% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{PowerMat}
\alias{PowerMat}
\title{Wavelet power matrix (from wavelet power cube)}
\usage{
PowerMat(x, ZScore = FALSE)
}
\arguments{
\item{x}{A cube with power matrices each slice representing ERP.}
}
\value{
An average power matrix (raw or as z-score).
}
\description{
This function computes the average power of several WT which can be returned as raw power
or as z-score. If the z-score is returned then every slice of the cube is z-transformed 
independently and the average is calculated in z-direction.
}
