% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{spike_ccf_batch}
\alias{spike_ccf_batch}
\title{Spike cross-correlation wrapper function for batch use}
\usage{
spike_ccf_batch(Time, UnitNr, WINDOW_LENGTH, BIN_SIZE)
}
\arguments{
\item{Time}{A numeric vector of times sorted in ascending order containing all time points.}

\item{UnitNr}{An integer vector containing the unit numbers in order of time occurence.}

\item{WINDOW_LENGTH}{A double as shape parameter determining the strength of centre exclusion.}

\item{BIN_SIZE}{A double as length parameter for gaussian kernel (2*KERNELSIZE+1).}
}
\value{
Returns a list containing the unit number, the counts per bin and the random expected count.
}
\description{
Computes the confidence interval of a poisson train using an inverted gaussian
distribution. Based on the weighted average with the given distribution lamda
can be estimated. The confidence interval is then estimated by estimation through
Chisq-distribution.
}
