% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{spike_stim_properties}
\alias{spike_stim_properties}
\title{Spike stimulation properties}
\usage{
spike_stim_properties(
  spike_idx,
  stim_mat_org,
  block_mat_org,
  sampling_rate,
  include_isolated
)
}
\arguments{
\item{spike_idx}{An integer vector representing the spike time point as sampling point (index).}

\item{stim_mat_org}{A matrix with stimulation properties (from Pulse_Seq_correctrion).}

\item{block_mat_org}{A matrix with stimulation block properties..}

\item{sampling_rate}{An integer indicating the sampling frequency.}

\item{include_isolated}{A bool indicting whether isolated stimulation outside of blocks should be included.}
}
\value{
Returns a matrix with stimulation properties for every spike ("onset", "peak_loc_offset", "peak_amp", "stim_length", "phase", "pulse", "sine", "ramp_front", "ramp_end", "frequency", "burst stim", "pulse_nr_block", "pulse_nr", "stimulation_block", "stimulation_timing", "pre", "post", "hyper_block", "hyper_block_frequency").
}
\description{
This function assigns stimulation properties to every spike so that it can be incorporated in a table.
}
