% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SpikeCCF}
\alias{SpikeCCF}
\title{Spike cross-correlation function

The function will calculate the cross-correlation for spikes of the same or a different cell. 
The input arguments can be a vector containg timpoints (in seconds), two different vectors 
with timepoints or one vector with time points for all units and an integer vector containg 
the unit ID. The ouput will be depending on the input a single CCF or a list containing the 
CCF for all combinations in a matrix. Additional options include the computation of the confidence 
interval and the baseline activity.}
\usage{
SpikeCCF(
  x,
  y = NULL,
  UnitNr = NULL,
  WINDOW_LENGTH = 1,
  BIN_SIZE = 0.001,
  BaselineFrequency = TRUE,
  ConfidenceInterval = TRUE,
  ConfLevel = 0.95,
  SD = 0.6,
  CENTREMIN = 0.6,
  KERNELSIZE = 20L
)
}
\arguments{
\item{x}{A numeric vector of times which has to be sorted in ascending order.}

\item{y}{A numeric vector of times which has to be sorted in ascending order.}

\item{UnitNr}{An integer vector containing the unit numbers in order of time occurence.}

\item{WINDOW_LENGTH}{An int as total window length in seconds (default = 1).}

\item{BIN_SIZE}{A double indicating the size of bins in seconds (default = 0.001).}

\item{BaselineFrequency}{A bool to indicate whether base line activity should be estimated (default = TRUE).}

\item{ConfidenceInterval}{A bool to indicate whether confidence interval should be estimated (default = TRUE).}

\item{ConfLevel}{A double indicating the confidence-level (default = 0.95).}

\item{SD}{A double as standard deviation for a gaussian shape parameter (default = 0.6).}

\item{CENTREMIN}{A double as shape parameter determining the strength of centre exclusion (default = 0.6).}

\item{KERNELSIZE}{A double as length parameter for gaussian kernel (2*KERNELSIZE+1, default = 20).}
}
\value{
Returns a list containing counts per bin, axis, random bin count, confidence-intervals with counts per bin.
}
\description{
Spike cross-correlation function

The function will calculate the cross-correlation for spikes of the same or a different cell. 
The input arguments can be a vector containg timpoints (in seconds), two different vectors 
with timepoints or one vector with time points for all units and an integer vector containg 
the unit ID. The ouput will be depending on the input a single CCF or a list containing the 
CCF for all combinations in a matrix. Additional options include the computation of the confidence 
interval and the baseline activity.
}
