% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SpikeCut}
\alias{SpikeCut}
\title{Spike cut out}
\usage{
SpikeCut(AmpMatrix, SpikeIdx, WINDOW = 20L)
}
\arguments{
\item{AmpMatrix}{A matrix with signal from the amplifier (rows = channels, columns = time).}

\item{SpikeIdx}{A matrix with signal from the amplifier (rows = channels, columns = time).}

\item{WINDOW}{An integer indicating the edges (default = 20).}
}
\value{
Returns a vector with position of maximum amplitude and amplitude itself.
}
\description{
This function extracts spike shape matrices from an ampliflier matrix (rows = channels, columns = time).
}
