% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SpikeMed}
\alias{SpikeMed}
\title{Spike median matrix}
\usage{
SpikeMed(SpikeCube)
}
\arguments{
\item{SpikeCube}{A cube containing matrices with isolated spikes (rows = channels, columns = time).}
}
\value{
Returns a matrix with median spike shapes (rows = channels, columns = time).
}
\description{
This function computes the median shape of spikes from a single spike cube.
}
