% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Squeeze}
\alias{Squeeze}
\title{Synchrosqueezed wavelet power matrix (from wavelet power matrix)}
\usage{
Squeeze(WT, frequencies, samplingfrequency = 1000, sigma = 6, CORES = 1L)
}
\arguments{
\item{WT}{A complex matrix representing the wavelet transform.}

\item{frequencies}{A vector indicating the frequencies which should be analysed.}

\item{samplingfrequency}{A double indicating the sampling frequency in Hz (default = 1000).}

\item{sigma}{A double indicating the shape parameter of the wavelet (default = 6).}

\item{CORES}{An integer indicating number of threads used (default = 1).}
}
\value{
A complex matrix representing the synchrosqueezed wavelet transform.
}
\description{
This function computes the synchrosqueezed wavelet transform as proposed by Daubechies and Maes (1996). Wavelet 
coefficients of the wavelet will be reassigned according to the instantaneous frequency in the transform.
}
