% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{StimFileRead}
\alias{StimFileRead}
\title{Load stimulation file}
\usage{
StimFileRead(FILENAME, digital = FALSE)
}
\arguments{
\item{FILENAME}{A string as path to file with single "/".}

\item{digital}{A bool indicating whether the file is a digital or analogin input (default = false).}
}
\value{
Returns an armadillo cube with extracted spikes.
}
\description{
This function extracts data points from a binary file either the analogin.dat or digitalin.dat generated intan system.
}
