% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{StimulusSequence}
\alias{StimulusSequence}
\title{Stimulus sequence}
\usage{
StimulusSequence(raw, sampling_frequency, threshold, max_time_gap)
}
\arguments{
\item{raw}{A numeric vector which is the stimulation over time as continues series.}

\item{sampling_frequency}{An integer indicating the sampling frequency.}

\item{threshold}{A double indicating the threshold of stimulus detection.}

\item{max_time_gap}{A double indicating the maximum time between blocks.}
}
\value{
Returns a list with a matrix showing single pulse properperties and a matrix with block properties.
}
\description{
This function analyses stimulus time series and extracts features which 
are returned in a list. Part of the output is the single stimulus feature
and the second part is a stimulus block summary.
}
