% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{UnitChannel}
\alias{UnitChannel}
\title{Spike location channel}
\usage{
UnitChannel(SpikeIdx, Units, AmpMatrix, WINDOW = 20L)
}
\arguments{
\item{SpikeIdx}{A numeric vector with spike indices.}

\item{Units}{A numeric vector unit numbers.}

\item{AmpMatrix}{A matrix with signal from the amplifier (rows = channels, columns = time).}
}
\value{
Returns a list inlcuding the channel number, amplitude and unit number.
}
\description{
This function is a wrapper to extract the channel location for any given unit based on the spike timings. It uses the spike 
indices, unit numbers, and the amplifier matrix (row = channels, Columns = time). It avoids using an apply function and 
directly accesses the data matrix.
}
