% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpikeExtractionMap.R
\name{UnitChannelMap}
\alias{UnitChannelMap}
\title{Get channel map}
\usage{
UnitChannelMap(
  SpikeIdx,
  Units,
  FileLoc,
  SubSampleSize,
  WindowSize,
  ChannelCount
)
}
\arguments{
\item{SpikeIdx}{Indeces for spike times.}

\item{Units}{Vector having units.}

\item{FileLoc}{String with file name}

\item{SubSampleSize}{Subsample of all spikes.}

\item{WindowSize}{Size of the window to extract spikes.}

\item{ChannelCount}{Number of Channels in file.}
}
\value{
Returns a data frame containing the coordinates of the channels.
}
\description{
Function wrapper to load channel map from a given directory.
}
