% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{WTCoherence}
\alias{WTCoherence}
\title{Smoothing complex wavelet matrix}
\usage{
WTCoherence(
  WT1,
  WT2,
  frequencies,
  samplingfrequency = 1000,
  sigma = 6,
  Ba = 0.6,
  Bb = 1
)
}
\arguments{
\item{WT1}{A complex matrix representing the wavelet transform.}

\item{WT2}{A complex matrix representing the wavelet transform.}

\item{frequencies}{A vector indicating the frequencies which should be analysed.}

\item{samplingfrequency}{A double indicating the sampling frequency in Hz (default = 1000).}

\item{sigma}{A double indicating the shape parameter of the wavelet (default = 6).}

\item{Ba}{A double indicating the smoothing factor in the scale domain (default = 0.6).}

\item{Bb}{A double indicating the smoothing factor in the time domain (default = 1).}
}
\value{
An smoothened wavelet transform.
}
\description{
This function computes the smoothened wavelet transform required 
for coherence calculation.
}
