% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ERPList}
\alias{ERPList}
\title{Wrapper for ERP (event-related potential) extraction

This function returns a list which includes the frequencies and matrices 
with extracted traces for any given range. The range is has to be provided
in form of onset indeces and end indeces. Additional information can be 
provided to adjust for sampling differences between range points and trace 
sampling frequency. If required the window is proportionally elongated to 
include Pre/Post times.}
\usage{
ERPList(
  Trace,
  BlockMat,
  SamplingFreqStim = 0,
  SamplingFreqTrace = 0,
  PrePhase = TRUE,
  PostPhase = TRUE,
  FixStartLength = 0,
  WindowLength = 0
)
}
\arguments{
\item{Trace}{A numeric vector which is used for extraction.}

\item{BlockMat}{The stimulation block matrix from StimulusSequence() including onset and end timings.}

\item{SamplingFreqStim}{A double indicating the sampling frequency of onset/end indeces.}

\item{SamplingFreqTrace}{A double indicating the sampling frequency of the trace.}

\item{PrePhase}{A bool indicating if pre-onset timings should be included (default = true).}

\item{PostPhase}{A bool indicating if post-end timings should be included (default = true).}

\item{FixStartLength}{A double indicating the starting length when fixed window is used (default = 0).}

\item{WindowLength}{A double indicating the time block in seconds (default = 0).}
}
\value{
Returns a list which includes stimulation frequencies for hyperblocks and the corresponding ERP matrices.
}
\description{
Wrapper for ERP (event-related potential) extraction

This function returns a list which includes the frequencies and matrices 
with extracted traces for any given range. The range is has to be provided
in form of onset indeces and end indeces. Additional information can be 
provided to adjust for sampling differences between range points and trace 
sampling frequency. If required the window is proportionally elongated to 
include Pre/Post times.
}
