% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{FirFiltering}
\alias{FirFiltering}
\title{FIR filtering}
\usage{
FirFiltering(SIGNAL, FIR_FILTER)
}
\arguments{
\item{SIGNAL}{A numeric vector.}

\item{FIR_FILTER}{A numeric vector which can be used as FIR filter.}
}
\value{
Returns numeric vector which is the FIR filtered original signal.
}
\description{
This function applies an FIR filter to a signal an returns the filtered trace.
}
