% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetChanMap.R
\name{loadChanMap}
\alias{loadChanMap}
\title{Load channel map
Function to load channel}
\usage{
loadChanMap(FileName)
}
\arguments{
\item{FileName}{Complete filename including path and file ending.}
}
\value{

}
\description{
Load channel map
Function to load channel
}
