% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaData.R
\name{MetaData}
\alias{MetaData}
\title{Extract meta data from ND-Manager xml file}
\usage{
MetaData(FileName)
}
\arguments{
\item{FileName}{String to input directory including complete file name.}
}
\value{
Returns a list which includes the number of channels, sampling rate, number of spike groups, and spike groups.
}
\description{
Function to extract meta data from xml file provided by the ND-Manager. Includes channel number,
sampling rate, spike groups.
}
