% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MITable.R
\name{MIPlot}
\alias{MIPlot}
\title{Plot modulation index from table}
\usage{
MIPlot(MITab, Colour = "magma", VLine = NULL, HLine = NULL)
}
\arguments{
\item{MITab}{data.table produced by MITable.}

\item{Colour}{Colour string based on viridis lookup table.}

\item{VLine}{Optional double for plotting vertical line.}

\item{HLine}{Optional double for plotting horizontal line.}
}
\value{
Returns a ggplot of the modulation index.
}
\description{
Plot modulation index from table
}
