% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{PhaseListAnalysis}
\alias{PhaseListAnalysis}
\title{Phase lock analysis}
\usage{
PhaseListAnalysis(x, CORES = 1L)
}
\arguments{
\item{x}{A cube of phases (radians) with slices as different ERPs.}

\item{CORES}{An int indicating the number of threads used (default = 1).}
}
\value{
Returns a list containing a matrix with Rho vector lengths and a matrix with the corresponding circular mean.
}
\description{
This function computes the phase lock using the Rho vector length 
(strength of the locking) and the circular mean phase from a list containing 
wavelet transforms of epochs. The input list has to be flattend (vectorised) 
first.
}
