% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{PhaseListAnalysisShuffle}
\alias{PhaseListAnalysisShuffle}
\title{Matrix shuffle}
\usage{
PhaseListAnalysisShuffle(x, DIM_X, DIM_Y, SHUFFLES, CORES)
}
\arguments{
\item{x}{A matrix.}

\item{DIM_X}{An int for x-dimension of matrix.}

\item{DIM_Y}{An int for y-dimension of matrix.}

\item{SHUFFLES}{An int indicating the number of shuffles.}

\item{CORES}{An int indicating the number of threads used (default = 1).}
}
\value{
Returns a matrix indicating the probability of value being larger than shuffled data.
}
\description{
This function shuffles the data independently of dimensions without 
resampling from a input matrix. It returns the probability for a any 
given value in a matrix to be larger than a random sample.
}
