% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{spike_ccf}
\alias{spike_ccf}
\title{Spike cross-correlation}
\usage{
spike_ccf(x, y, WINDOW_LENGTH = 1, BIN_SIZE = 0.001)
}
\arguments{
\item{x}{A numeric vector of times which has to be sorted in ascending order.}

\item{y}{A numeric vector of times which has to be sorted in ascending order.}

\item{WINDOW_LENGTH}{An int as total window length in seconds (default = 1).}

\item{BIN_SIZE}{A double indicating the size of bins in seconds (default = 0.001).}
}
\value{
Integer vector with counts per bin.
}
\description{
This function computes the cross-correlation for spike trains.
In case of a cross-correlation of the same time series an
auto-correlation will be computed and with 0 correlation correction.
}
