% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SpikeCCF}
\alias{SpikeCCF}
\title{Spike cross-correlation wrapper function}
\usage{
SpikeCCF(
  x,
  y,
  WINDOW_LENGTH = 1,
  BIN_SIZE = 0.001,
  BaselineFrequency = TRUE,
  ConfidenceInterval = TRUE,
  ConfLevel = 0.95,
  SD = 0.6,
  CENTREMIN = 0.6,
  KERNELSIZE = 20L
)
}
\arguments{
\item{x}{A numeric vector of times which has to be sorted in ascending order.}

\item{y}{A numeric vector of times which has to be sorted in ascending order.}

\item{WINDOW_LENGTH}{An int as total window length in seconds (default = 1).}

\item{BIN_SIZE}{A double indicating the size of bins in seconds (default = 0.001).}

\item{BaselineFrequency}{A bool to indicate whether base line activity should be estimated (default = TRUE).}

\item{ConfidenceInterval}{A bool to indicate whether confidence interval should be estimated (default = TRUE).}

\item{ConfLevel}{A double indicating the confidence-level (default = 0.95).}

\item{SD}{A double as standard deviation for a gaussian shape parameter (default = 0.6).}

\item{CENTREMIN}{A double as shape parameter determining the strength of centre exclusion (default = 0.6).}

\item{KERNELSIZE}{A double as length parameter for gaussian kernel (2*KERNELSIZE+1, default = 20).}
}
\value{
Returns a list containing counts per bin, axis, random bin count, confidence-intervals with counts per bin.
}
\description{
This function computes the cross-correlation for spike trains.
In case of a cross-correlation of the same time series an auto-correlation 
will be computed and with 0 correlation correction.
The function has the capability to output the corresponding confidence-
interval based on an inverse gaussian lamda approximation. Further a random
distribution of spike occurence is computed based on spike number and time.
}
