% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SpikeCCF}
\alias{SpikeCCF}
\title{Spike cross-correlation wrapper function}
\usage{
SpikeCCF(
  x,
  y,
  WINDOW_LENGTH = 1,
  BIN_SIZE = 0.001,
  BaselineFrequency = TRUE,
  ConfidenceInterval = TRUE,
  ConfLevel = 0.95,
  SD = 0.6,
  CENTREMIN = 0.6,
  KERNELSIZE = 20L
)
}
\arguments{
\item{x}{A numeric vector of times which has to be sorted in ascending order.}

\item{y}{A numeric vector of times which has to be sorted in ascending order.}

\item{WINDOW_LENGTH}{An int as filtering order (default = 2).}

\item{BIN_SIZE}{A double as cutoff frequency.}
}
\value{
Returns a list containing counts per bin, axis, random bin count, confidence-intervals with counts per bin.
}
\description{
This function computes the cross-correlation for spike trains.
In case of a cross-correlation of the same time series an auto-
correlation will be computed and with 0 correlation correction.
The function has the capability to output the corresponding confidence-
interval based on an inverse gaussian lamda approximation. Further a random
distribution of spike occurence is computed based on spike number and time.
}
