% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{WT}
\alias{WT}
\title{Wavelet transform}
\usage{
WT(Signal, frequencies, samplingfrequency, sigma, LNorm = 2, CORES = 1L)
}
\arguments{
\item{Signal}{A numeric vector.}

\item{frequencies}{A vector indicating the frequencies which should be analysed.}

\item{samplingfrequency}{A double indicating the sampling frequency in Hz.}

\item{sigma}{A double indicating the shape parameter of the wavelet.}

\item{LNorm}{A double indicating the L normalisation (power of 1/LNorm, default = 2).}

\item{CORES}{An integer indicating number of threads used (default = 1).}
}
\value{
Wavelet transform as complex matrix.
}
\description{
This function performs a wavelet transform of a signal for different scales 
and returns a complex matrix from the convolution with a complex wavelet in 
the frequency domain.
}
