% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{WTbatch}
\alias{WTbatch}
\title{Wavelet transform (from ERP matrix)

This function performs a wavelet transform of a signal for different scales 
and returns a complex matrix from the convolution with a complex wavelet in 
the frequency domain.}
\usage{
WTbatch(
  ERPMat,
  frequencies,
  SamplingRate = 1000,
  sigma = 6,
  LNorm = 2,
  CORES = 1L,
  compression = FALSE,
  PhaseAnalysis = FALSE
)
}
\arguments{
\item{ERPMat}{A numeric matrix with rows for ERP signals and columns as time domain.}

\item{frequencies}{A vector indicating the frequencies which should be analysed.}

\item{SamplingRate}{A double indicating the sampling frequency in Hz (default = 1000).}

\item{sigma}{A double indicating the shape parameter of the wavelet (default = 6).}

\item{LNorm}{A double indicating the L normalisation (power of 1/LNorm, default = 2).}

\item{CORES}{An integer indicating number of threads used (default = 1).}

\item{compression}{An integer indicating number of threads used (default = 1).}

\item{PhaseAnalysis}{An integer indicating number of threads used (default = 1)}
}
\value{
List with wavelet transform as complex cube (each slice is from one ERP or when compressed one matrix), 
rho vector length and mean phase.
}
\description{
Wavelet transform (from ERP matrix)

This function performs a wavelet transform of a signal for different scales 
and returns a complex matrix from the convolution with a complex wavelet in 
the frequency domain.
}
