% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{WTCoherence}
\alias{WTCoherence}
\title{Coherence/Coherency/Phase Difference function}
\usage{
WTCoherence(
  WT1,
  WT2,
  frequencies,
  SamplingRate = 1000,
  tKernelWidth = 0.01,
  sKernelWidth = 0.6
)
}
\arguments{
\item{WT1}{A complex matrix representing the wavelet transform.}

\item{WT2}{A complex matrix representing the wavelet transform.}

\item{frequencies}{A vector indicating the frequencies which should be analysed.}

\item{SamplingRate}{A double indicating the sampling frequency in Hz (default = 1000).}

\item{tKernelWidth}{A double indicating the sd as smoothing factor in the time domain (default = 0.01).}

\item{sKernelWidth}{A double indicating the smoothing factor in the scale domain (default = 0.6).}
}
\value{
A list containing the coherency, coherence, and phase difference of the two wavelet transforms.
}
\description{
This function computes the coherence, coherency, and the phase difference between two wavelet transforms.
}
